Add append test
- TestAppend
- Client.Append now returns file offset + error instead of just offset.
- Add error code for reporting append data size limit exceeded.
- Add error code for checking lease holder not found.
- Add constants in test_test.go for ease of testing.
- Client.pushData now returns error instead of bool.
- Bugfix: Correctly calculate when there is not enough space left for
  record appends.
- Bugfix: ChunkServer.Append now returns correct file offset, instead
  of returning chunk offset.
- Bugfix: Proper error code equality checking in
  Client.guaranteeChunkLocations for sgfsErr.
- Bugfix: ChunkManager.addChunk now returns sgfsErr.ErrChunkExist when
  adding an existing chunk instead of errors.New().

+++ common.go
@@ -157,7 +157,7 @@ type AppendArgs struct {
 }
 
 type AppendReply struct {
-  Offset int
+  Offset uint64
 }
 
 // Helper functions
